+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created TODO.md inspired by the 'Keep a Changelog' format
 - Add quiet setting and flag to suppress regular output
 - Add version flag to display current version number
+- Add simulate setting and flag to display what happens if run normally with no filesystem modifications
 
 ## [0.0.2] - 2022-08-28